CloudStore integration & UI changes

+++ app/src/main/java/com/example/matsnbeltsassociate/activity/InfoPopupActivity.java
@@ -15,7 +15,7 @@ public class InfoPopupActivity extends AppCompatActivity {
         getWindowManager().getDefaultDisplay().getMetrics(dm);
         int width = dm.widthPixels;
         int height = dm.heightPixels;
-        getWindow().setLayout((int) (width * 6), (int) (height * 5));
+        getWindow().setLayout((width * 6), (height * 5));
 
         WindowManager.LayoutParams params = getWindow().getAttributes();
         params.gravity = Gravity.CENTER;
